requirements.txt
earthengine-api>=0.1.390
requests>=2.31.0
Pillow>=10.0.0
rasterio>=1.3.9
numpy>=1.26.0
matplotlib>=3.8.0
scikit-image>=0.22.0
scipy>=1.11.0